move channel files into SimpleChatServer; remove some unused imports, edit spacing
